fix another bug

+++ setup-linux-2.sh
@@ -6,10 +6,15 @@ sudo apt-get install postgresql postgresql-contrib libpq-dev
 echo -e "CREATE USER samfundet WITH PASSWORD 'samfundet';\nALTER USER samfundet CREATEDB;" | sudo -u postgres psql
 bundle install
 
+if [ ! -e $CONFIG_DIR/database.yml ] &&
+   [ ! -e $CONFIG_DIR/local_env.yml ] &&
+   [ ! -e $CONFIG_DIR/billig.yml ] &&
+   [ ! -e $CONFIG_DIR/secrets.yml ]; then
+     make copy-config-files || exit
+fi
+
+rm config/database.yml
 cp config/database.example.yml config/database.yml
-cp config/local_env.example.yml config/local_env.yml
-cp config/billig.example.yml config/billig.yml
-cp config/secrets.example.yml config/secrets.yml
 
 sed -i "s/password:.*/password: samfundet/" > config/database.yml
